First commit! Basic struct of the test app.
This is a basic outline of the various parts of the node app. Note things here are conceptual at this point and do not represent final implementation decisions. The Discord Bot will start and listen to messages, however no logic has been provided to the command interface yet. Also no glue has been provided to dynamically load command classes. The API for the command interface is lightly modeled after the PHP symphony framework API, of course adapted to JS usecases.